remove flip buffer limits/pointers and share back buffer limits/pointers
This simplification is intended to make implementation of EGL_BUFFER_AGE_EXT
easier: The funtionality eglQuerySurface(..., EGL_BUFFER_AGE_EXT) must be
implemented in egl.c since there is no callback for eglQuerySurface in
waylandwsegl.c. To work properly we must ensure that the number of back buffers
is fixed in either flipping/blitting case.

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>